Refactor auth imports to use the core directory for consistency.
